crew-ai/pyproject.toml
[project]
name = "marketing_posts"
version = "0.1.0"
description = "marketing-posts using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.85.0",
    "crewai-tools[mcp]>=0.4.6",
    "setuptools (>=80.9.0,<81.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "types-pyyaml (>=6.0.12.20250516,<7.0.0.0)",
]

[project.scripts]
marketing_posts = "marketing_posts.main:run"
train = "marketing_posts.main:train"
run_crew = "marketing_posts.main:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "ruff>=0.12.0",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
pyright = "^1.1.402"

